Ensure assets are also included in error views.
Register webassets_injector_tween after excview_tween_factory so that
assets are included in error views.

Why this?
Here is a real world example to demonstrate why this is beneficial:
Say you have a 404 page with navigation and it uses `Webassets`,
the navigation will not actually work for all 404 pages.